체중, 운동입력, 기록 페이지 기능 구현

+++ src/main/java/com/example/demo/dto/user/UserDto.java
@@ -29,8 +29,8 @@ public class UserDto {
 	private int age; // 나이
 	private int height; // 키
 	private String activityLevel; // 활동계수 (Sedentary, Lightly Active, Moderately Active, Very Active)
-	private int currentWeight; // 현재 체중
-	private int goalWeight; // 목표 체중
+	private float currentWeight; // 현재 체중
+	private float goalWeight; // 목표 체중
 	private int cheatingDay; // 치팅 데이 (숫자로 표현된 값)
 	private String type; // 식단 계획 유형 (custom, common, health, quito, vegan)
 	private int carbohydrate; // 탄수화물
